Separate hook for getVisible slides per page
